requirements.txt
opencv-python
requests
RPi.GPIO
